modified cleanup function to include filter

+++ compute/cloud-client/src/test/java/compute/Util.java
@@ -16,9 +16,16 @@
 
 package compute;
 
+import com.google.cloud.compute.v1.AggregatedListInstancesRequest;
 import com.google.cloud.compute.v1.Instance;
+import com.google.cloud.compute.v1.Instance.Status;
 import com.google.cloud.compute.v1.InstanceTemplate;
+import com.google.cloud.compute.v1.InstanceTemplatesClient;
+import com.google.cloud.compute.v1.InstanceTemplatesClient.ListPagedResponse;
+import com.google.cloud.compute.v1.InstancesClient;
+import com.google.cloud.compute.v1.InstancesClient.AggregatedListPagedResponse;
 import com.google.cloud.compute.v1.InstancesScopedList;
+import com.google.cloud.compute.v1.ListInstanceTemplatesRequest;
 import java.io.IOException;
 import java.time.Instant;
 import java.time.OffsetDateTime;
@@ -38,13 +45,14 @@ public class Util {
   // has creation timestamp >24 hours.
   public static void cleanUpExistingInstanceTemplates(String prefixToDelete, String projectId)
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    for (InstanceTemplate template : ListInstanceTemplates.listInstanceTemplates(projectId)
+    for (InstanceTemplate template : listFilteredInstanceTemplates(projectId, prefixToDelete)
         .iterateAll()) {
       if (!template.hasCreationTimestamp()) {
         continue;
       }
       if (template.getName().contains(prefixToDelete)
-          && isCreatedBeforeThresholdTime(template.getCreationTimestamp())) {
+          && isCreatedBeforeThresholdTime(template.getCreationTimestamp())
+          && template.isInitialized()) {
         DeleteInstanceTemplate.deleteInstanceTemplate(projectId, template.getName());
       }
     }
@@ -56,14 +64,15 @@ public class Util {
   public static void cleanUpExistingInstances(String prefixToDelete, String projectId,
       String instanceZone)
       throws IOException, ExecutionException, InterruptedException, TimeoutException {
-    for (Entry<String, InstancesScopedList> instanceGroup : ListAllInstances.listAllInstances(
-        projectId).iterateAll()) {
+    for (Entry<String, InstancesScopedList> instanceGroup : listFilteredInstances(
+        projectId, prefixToDelete).iterateAll()) {
       for (Instance instance : instanceGroup.getValue().getInstancesList()) {
         if (!instance.hasCreationTimestamp()) {
           continue;
         }
         if (instance.getName().contains(prefixToDelete)
-            && isCreatedBeforeThresholdTime(instance.getCreationTimestamp())) {
+            && isCreatedBeforeThresholdTime(instance.getCreationTimestamp())
+            && instance.getStatus().equalsIgnoreCase(Status.RUNNING.toString())) {
           DeleteInstance.deleteInstance(projectId, instanceZone, instance.getName());
         }
       }
@@ -73,6 +82,34 @@ public class Util {
   public static boolean isCreatedBeforeThresholdTime(String timestamp) {
     return OffsetDateTime.parse(timestamp).toInstant()
         .isBefore(Instant.now().minus(DELETION_THRESHOLD_TIME_HOURS, ChronoUnit.HOURS));
+  }
+
+  public static AggregatedListPagedResponse listFilteredInstances(String project,
+      String instanceNamePrefix) throws IOException {
+    try (InstancesClient instancesClient = InstancesClient.create()) {
+
+      AggregatedListInstancesRequest aggregatedListInstancesRequest = AggregatedListInstancesRequest
+          .newBuilder()
+          .setProject(project)
+          .setFilter(String.format("name:%s", instanceNamePrefix))
+          .build();
+
+      return instancesClient
+          .aggregatedList(aggregatedListInstancesRequest);
+    }
+  }
+
+  public static ListPagedResponse listFilteredInstanceTemplates(String projectId,
+      String instanceTemplatePrefix) throws IOException {
+    try (InstanceTemplatesClient instanceTemplatesClient = InstanceTemplatesClient.create()) {
+      ListInstanceTemplatesRequest listInstanceTemplatesRequest =
+          ListInstanceTemplatesRequest.newBuilder()
+              .setProject(projectId)
+              .setFilter(String.format("name:%s", instanceTemplatePrefix))
+              .build();
+
+      return instanceTemplatesClient.list(listInstanceTemplatesRequest);
+    }
   }
 
 }
